+++ Packages/MacUI/Sources/ControlUI/SettingsWindow/SettingsWindow.swift
@@ -9,7 +9,7 @@
 //
 //  ---------------------------------------------------------------------------
 //
-//  © 2018-2024 1024jp
+//  © 2018-2025 1024jp
 //
 //  Licensed under the Apache License, Version 2.0 (the "License");
 //  you may not use this file except in compliance with the License.
@@ -26,9 +26,15 @@
 
 import AppKit
 
-final class SettingsWindow: NSPanel {
+final class SettingsWindow: NSWindow {
     
     // MARK: Window Methods
+    
+    override func cancelOperation(_ sender: Any?) {
+    
+        self.close()
+    }
+    
     
     override func validateMenuItem(_ menuItem: NSMenuItem) -> Bool {