+++ Packages/MacUI/Sources/ControlUI/SettingsWindow/SettingsWindowController.swift
@@ -9,7 +9,7 @@
 //
 //  ---------------------------------------------------------------------------
 //
-//  © 2023-2024 1024jp
+//  © 2023-2025 1024jp
 //
 //  Licensed under the Apache License, Version 2.0 (the "License");
 //  you may not use this file except in compliance with the License.
@@ -43,7 +43,6 @@ public final class SettingsWindowController<Pane: SettingsPane>: NSWindowControl
         
         let window = SettingsWindow(contentViewController: viewController)
         window.styleMask = [.closable, .titled]
-        window.hidesOnDeactivate = false
         
         super.init(window: window)
     }